requirements.txt
flask
pyserial


pyinstaller --onefile --windowed --icon=wricon.ico --name="RentalTide" scanner.py
